trade/requirements.txt
python-dotenv==0.10.1
pymongo==3.7.2
alpaca-trade-api==0.24
pandas>=0.23.4
